+++ src/Silf.cpp
@@ -199,8 +199,8 @@ bool Silf::readGraphite(const byte * const silf_start, size_t lSilf, Face& face,
     {
         uint32 pass_start = be::read<uint32>(o_passes);
         uint32 pass_end = be::peek<uint32>(o_passes);
-        face.error_context((face.error_context() & 0xFF00) + EC_ASILF + (i << 16));
-        if (e.test(pass_start > pass_end, E_BADPASSSTART) 
+        face.error_context((face.error_context() & 0xFF00) + EC_ASILF + unsigned(i << 16));
+        if (e.test(pass_start > pass_end, E_BADPASSSTART)
                 || e.test(pass_start < passes_start, E_BADPASSSTART)
                 || e.test(pass_end > lSilf, E_BADPASSEND)) {
             releaseBuffers(); return face.error(e);
@@ -233,7 +233,7 @@ bool Silf::readGraphite(const byte * const silf_start, size_t lSilf, Face& face,
 template<typename T> inline uint32 Silf::readClassOffsets(const byte *&p, size_t data_len, Error &e)
 {
     const T cls_off = 2*sizeof(uint16) + sizeof(T)*(m_nClass+1);
-    const size_t max_off = (be::peek<T>(p + sizeof(T)*m_nClass) - cls_off)/sizeof(uint16);
+    const uint32 max_off = (be::peek<T>(p + sizeof(T)*m_nClass) - cls_off)/sizeof(uint16);
     // Check that the last+1 offset is less than or equal to the class map length.
     if (e.test(be::peek<T>(p) != cls_off, E_MISALIGNEDCLASSES)
             || e.test(max_off > (data_len - cls_off)/sizeof(uint16), E_HIGHCLASSOFFSET))
@@ -357,7 +357,7 @@ uint16 Silf::getClassGlyph(uint16 cid, unsigned int index) const
 bool Silf::runGraphite(Segment *seg, uint8 firstPass, uint8 lastPass, int dobidi) const
 {
     assert(seg != 0);
-    unsigned int       maxSize = seg->slotCount() * MAX_SEG_GROWTH_FACTOR;
+    size_t             maxSize = seg->slotCount() * MAX_SEG_GROWTH_FACTOR;
     SlotMap            map(*seg, m_dir, maxSize);
     FiniteStateMachine fsm(map, seg->getFace()->logger());
     vm::Machine        m(map);
